requirements.txt
discord.py==2.3.2 
mysql-connector==2.2.9
mysqlclient==2.2.1 
openai==0.28.1
pandas==2.1.4
py-cord==2.0.0b7
requests==2.31.0
